Moved components stories into stories folder
