+++ .storybook/main.js
@@ -15,6 +15,10 @@ module.exports = {
       use: ['style-loader', 'css-loader', 'sass-loader'],
       include: path.resolve(__dirname, '../'),
     });
+
+    config.resolve.alias = {
+      "@": path.resolve(__dirname, "..", "src")
+    }
   
     return config;
   }
