+++ src/stories/components/FileCard.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, ArgsTable, Preview } from '@storybook/addon-docs/blocks';
 
-import FileCard from '../FileCard/FileCard'
+import FileCard from '../../components/FileCard/FileCard'
 
 <!-- Story args -->
 <!-- https://storybook.js.org/docs/react/writing-stories/args -->
