+++ src/stories/components/ProgressBar.stories.js
@@ -1,4 +1,4 @@
-import ProgressBar from '../ProgressBar'
+import ProgressBar from '../../components/ProgressBar'
 
 
 export default {
